fix tests

+++ services/s3/src/main/java/software/amazon/awssdk/services/s3/internal/multipart/DownloadObjectHelper.java
@@ -31,7 +31,7 @@ public class DownloadObjectHelper {
 
     private final S3AsyncClient s3AsyncClient;
     private final long bufferSizeInBytes;
-    private final long maxInFlight;
+    private final int maxInFlight;
 
     public DownloadObjectHelper(S3AsyncClient s3AsyncClient, long bufferSizeInBytes, int maxInFlight) {
         this.s3AsyncClient = s3AsyncClient;
